backend/blueprints/ml_data_pipeline/requirements.txt
sagemaker==2.57.0
tabulate==0.8.9
pyyaml==5.4.1
jsonschema==3.2.0
jinja2==3.0.1
toposort==1.6
aws-embedded-metrics==1.0.7
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
pytest==6.2.5
boto3==1.18.33
boto3-stubs==1.18.33
aws-cdk.core==1.132.0
aws-cdk.aws-lambda==1.132.0
aws-cdk.aws-lambda-python==1.132.0
aws-cdk.aws-lambda-event-sources==1.132.0
aws-cdk.aws-events==1.132.0
aws-cdk.aws-cognito==1.132.0
aws-cdk.aws-rds==1.132.0
aws-cdk.aws-s3==1.132.0
aws-cdk.aws_s3_deployment==1.132.0
aws-cdk.aws_sagemaker==1.132.0
aws-cdk.aws-glue==1.132.0
aws-cdk.aws-codecommit==1.132.0
aws-cdk.aws-codebuild==1.132.0
aws-cdk.aws-codepipeline==1.132.0
aws-cdk.pipelines==1.132.0
aws-cdk.aws_apigateway==1.132.0
aws-cdk.aws_ecs_patterns==1.132.0
aws-cdk.aws-cloudfront-origins==1.132.0
aws-cdk.aws_stepfunctions_tasks==1.132.0
aws-cdk.aws_sns==1.132.0
aws-cdk.aws_dynamodb==1.132.0
aws-cdk.aws_athena==1.132.0
aws-cdk.aws-batch==1.132.0
aws-cdk.aws_iam==1.132.0
pyspark==3.1.2
nanoid==2.0.0
structlog==21.1.0
psutil==5.8.0